requirements.txt
# Telegram Bot
aiogram>=3.3.0

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# HTTP Requests
aiohttp>=3.9.1

# AI / OpenAI
openai>=1.12.0

# Scheduler for reminders
apscheduler>=3.10.4


python-dateutil>=2.8.2
pytz>=2024.1

# Logging
python-dotenv>=1.0.1

# SQLite async support
aiosqlite>=0.19.0

# Email Integration (IMAP)
imapclient>=3.0.1
beautifulsoup4>=4.12.3
lxml>=5.1.0
cryptography>=42.0.0
email-validator>=2.1.0

# Document Processing (attachments)
openpyxl>=3.1.2
python-docx>=1.1.0
xlrd>=2.0.1
